# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Orbit Copilot API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chat/messages:
    get:
      operationId: getChatMessages
      tags: [chat]
      summary: Get chat history
      responses:
        "200":
          description: List of chat messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatMessage"
    post:
      operationId: sendChatMessage
      tags: [chat]
      summary: Send a message to the AI copilot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMessageInput"
      responses:
        "201":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatMessage"
  /chat/messages/{id}:
    delete:
      operationId: deleteChatMessage
      tags: [chat]
      summary: Delete a chat message
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ChatMessage:
      type: object
      properties:
        id:
          type: integer
        role:
          type: string
          enum: [user, assistant]
        content:
          type: string
        createdAt:
          type: string
        metadata:
          type: ["object", "null"]
      required: [id, role, content, createdAt]
    ChatMessageInput:
      type: object
      properties:
        content:
          type: string
          minLength: 1
        context:
          type: ["string", "null"]
      required: [content]
...